Updated their import statements to import the renamed sequelize connection instance
