+++ server/models/Transaction.js
@@ -1,5 +1,5 @@
 const {DataTypes, Model } = require('sequelize');
-const {sequelize, connectDB} = require('../config/db');
+const {sequelize, connectDB} = require('../config/connection');
 const Account = require('./Account');
 const RecurrTransacGrp = require('./RecurrTransacGrp');
